% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cole-package.R
\docType{package}
\name{cole-package}
\alias{cole}
\alias{cole-package}
\title{cole: COmpound Loss Emporium}
\description{
Procedures for minimizing various compound loss functions, e.g. squared error loss for the simultaneous estimation of normal means.
}
\author{
\strong{Maintainer}: Dave Zhao \email{dave.zhao@gmail.com}

}
