% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comte_min}
\alias{comte_min}
\title{Compound decision method for two-sieded models with minimal sigma square}
\usage{
comte_min(y, x, ss, B, tol, maxit, p, q, n, newx)
}
\arguments{
\item{y}{outcomes for training with dimension n by q}

\item{x}{features for training with dimension n by p}

\item{ss}{condidates sigma square estimated by other mithods (lasso, ridge, camel)}

\item{B}{condidates beta estimated by other mithods (lasso, ridge, camel)}

\item{tol}{error tolerance for convergence}

\item{maxit}{maximum number of allowable iterations}

\item{p}{number of features}

\item{q}{number of outcomes}

\item{n}{number of observations for training}

\item{newx}{new observation for prediction}
}
\value{
\item{f}{estimation of prior of beta}
\item{A}{likelihood of training data}
\item{bs}{candidates beta}
\item{esty}{estimation for based on newx}
}
\description{
EM algorithm for predicting unseen data for two-sided linear model based on candidate parameters estimated by simple models. Sigma square is taken to be the minimum for better accuracy in some situations
}
\examples{
\donttest{
p = 10
q = 5
n = 50

x = matrix(rnorm(n*p,0,10), n, p)
beta = matrix(rnorm(p*q,0,10), q, p)
e = matrix(rnorm(n*q,0,0.1),n,q)
y = x \%*\% t(beta) + e
s2 = matrix(rep(0.1,q), q, 1)
tol = 0.001
maxit = 1000
x_test = matrix(rnorm(n*p,0,1), n, p)

output = comte_min(y, x, s2, beta, tol, maxit, p, q, n, x_test)

}

}
