% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolish.R
\name{coolish_con}
\alias{coolish_con}
\title{Constrained Coordinate Linear Shrinkage method (COOLISH)}
\usage{
coolish_con(x_tr, y_tr, x_te)
}
\arguments{
\item{x_tr}{training predictor matrix (n_tr x p)}

\item{y_tr}{training response matrix (n_tr x q)}

\item{x_te}{testing predictor matrix (n_te x p)}
}
\value{
\item{y}{estimated response matrix of testing set (n_te x q)}
}
\description{
Linearly shrinks coordinate of least squared estimates with constrains to solve multivariate linear prediction problems.
}
\examples{
n_tr = 100
n_te = 10
p = 30
q = 1000

set.seed(1)
B = matrix(rnorm(p*q), p, q)
B[6:p, ] = 0
x_tr = matrix(rnorm(n_tr*p), n_tr, p)
x_te = matrix(rnorm(n_te*p), n_te, p)

y_tr = x_tr \%*\% B + matrix(rnorm(n_tr*q), n_tr, q)
y_te = x_te \%*\% B + matrix(rnorm(n_te*q), n_te, q)

y_coolish = coolish_con(x_tr, y_tr, x_te)
fit_ols = lm(y_tr~x_tr)
y_ols = cbind(1, x_te) \%*\% fit_ols$coefficients

mean((y_te-y_coolish)^2)
mean((y_te-y_ols)^2)
}
