% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf1.R
\name{cv.wtf1}
\alias{cv.wtf1}
\title{Monotone Degree-One Weighted Trend Filter Cross-Validation}
\usage{
cv.wtf1(x, s, tau.options, nfolds = 5, seed = NULL, fit.best = TRUE, ...)
}
\arguments{
\item{x}{Gaussian sequence}

\item{s}{standard deviation}

\item{tau.options}{a vector of candidate values for TV(d')}

\item{nfolds}{number of cross-validation folds}

\item{seed}{random seed used for reproducible folds}

\item{fit.best}{boolean (default TRUE) whether to fit all of \code{x} with the
best candidate value of TV(d')}

\item{...}{additional parameters passed to wtf1() such as backend and knots}
}
\value{
\item{theta_hat}{estimated values of means of Gaussian sequence}
\item{x}{original Gaussian sequence}
\item{s}{known standard deviation}
\item{SURE}{value of minimum risk estimate}
\item{tau}{the user specified bound on TV(d')}
\item{TV1}{the value of TV(d') for minimizer}
\item{intercept}{intercept of fitted estimator}
\item{slopes}{slopes of fitted estimator}
\item{knots}{location of knots}
\item{backend}{which version of wtf1 was used}
\item{solver}{solver used by CVXR}
\item{...}{additional inputed parameters}
\item{cv.risks}{data.frame containing cross-validation parameters and risks}
\item{nfold}{number of folds for cross-validation}
\item{seed}{seed for reproducible fold partitions}
}
\description{
Use a fixed set of knots and a random partition of the data to cross-validate
the risk estimate for a provided list of tau values. If \code{nfolds = length(x)}
leave-one-out cross-validation is used. If \code{fit.best = TRUE} (default) then
a final model is fit using all of the data and best value of tau. A seed is
available to set for reproducible partitions.
}
\examples{
# basic usage
set.seed(1)
theta = rnorm(250)
x = theta + rnorm(250)
res = cv.wtf1(x, s = 1, tau.options = 1:2, knots = -2:2)
mean((theta - res$theta_hat)^2)
}
