% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_kde.R
\name{f.kde}
\alias{f.kde}
\title{Kernel Density f-Estimator}
\usage{
f.kde(x, s, bw = s/sqrt(log(length(x))), Cn = s * sqrt(3 * log(length(x))))
}
\arguments{
\item{x}{Gaussian sequence}

\item{s}{standard deviation}

\item{bw}{scalar bandwidth for Gaussian kernel density estimate}

\item{Cn}{scalar threshold for to clip estimates in low-density areas (Inf for no thresholding)}
}
\value{
\item{theta_hat}{estimated values of means of Gaussian sequence}
}
\description{
An implementation of the f-estimator from "Nonparametric empirical Bayes and
compound decision approaches to estimation of a high-dimensional vector of
normal means" (Brown and Greenshtein, 2009) for estimating the mean of a
homoscedastic sequence of independent Gaussian observations under squared
error loss. The various small modifications and default parameters that are
suggested in their paper are implemented.
}
\examples{
theta = rnorm(250)
x = theta + rnorm(250)
res = f.kde(x, s = 1)
mean((theta - res$theta_hat)^2)
}
