% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g_rule}
\alias{g_rule}
\title{Separable rule for Gaussian mean estimation}
\usage{
g_rule(x, s, t, rho = 0)
}
\arguments{
\item{x}{primary Gaussian sequence}

\item{s}{standard deviation of primary sequence}

\item{t}{tuning parameter vector t1}

\item{rho}{regularization parameter, closer to 0 means less regularization}
}
\value{
estimated values of means of primary Gaussian sequence
}
\description{
Given tuning parameter vector t, returns corresponding estimate for the mean vector of a homoscedastic sequence of independent Gaussian observations
}
\examples{
## generate data
n = 250
set.seed(1)
theta = rnorm(n)
x = theta + rnorm(n)
## loss of MLE
mean((theta - x)^2)
## loss of oracle separable estimator
mean((theta - g_rule(x, 1, theta))^2)
}
