% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg}
\alias{gg}
\title{Gaussian mean estimation with Gaussian side information}
\usage{
gg(x1, s1, x2, s2, rho = 0, K = 10, C = 5, tol = 1e-05,
  maxit = 100)
}
\arguments{
\item{x1}{primary Gaussian sequence}

\item{s1}{standard deviation of primary sequence}

\item{x2}{auxiliary Gaussian sequence of side information}

\item{s2}{standard deviation of auxiliary sequence}

\item{rho}{regularization parameter, closer to 0 means less regularization}

\item{K}{number of grid points, in the interval [x_j1 - C s_1, x_j1 + C s1], over which to search for the jth tuning parameter, where C is define below}

\item{C}{the value of the constant C in the interval above}

\item{tol}{error tolerance for convergence of the unbiased risk estimate}

\item{maxit}{maximum number of allowable iterations}
}
\value{
\item{t1_hat}{values of tuning parameters t1}
\item{t2_hat}{values of tuning parameters t2}
\item{theta_hat}{estimated values of means of primary Gaussian sequence}
}
\description{
Estimates the mean vector of a primary homoscedastic sequence of independent Gaussian observations under squared error loss by leveraging side information in the form of an auxiliary homoscedastic sequence of independent Gaussians.
}
\examples{
\donttest{
## generate data
n = 250
set.seed(1)
theta1 = rnorm(n)
theta2 = theta1
x1 = theta1 + rnorm(n)
x2 = theta2 + rnorm(n)
## loss of MLE
mean((theta1 - x1)^2)
## loss of estimator incorporating side information
mean((theta1 - gg(x1, 1, x2, 1)$theta_hat)^2)
}

}
