% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_rule}
\alias{gg_rule}
\title{Separable rule for Gaussian mean estimation with Gaussian side information}
\usage{
gg_rule(x1, s1, x2, s2, t1, t2, r = 0, rho = 0)
}
\arguments{
\item{x1}{primary Gaussian sequence}

\item{s1}{standard deviation of primary sequence}

\item{x2}{auxiliary Gaussian sequence of side information}

\item{s2}{standard deviation of auxiliary sequence}

\item{t1}{tuning parameter vector t1}

\item{t2}{tuning parameter vector t2}

\item{r}{correlation between x_i1 and x_i2, assuming they are bivariate normal}

\item{rho}{regularization parameter, closer to 0 means less regularization}
}
\value{
estimated values of means of primary Gaussian sequence
}
\description{
Given tuning parameter vectors t1 and t2, returns corresponding estimate for the mean vector of a primary homoscedastic sequence of independent Gaussian observations that leverages side information in the form of an auxiliary homoscedastic sequence of independent Gaussians.
}
\examples{
## generate data
n = 250
set.seed(1)
theta1 = rnorm(n)
theta2 = theta1
x1 = theta1 + rnorm(n)
x2 = theta2 + rnorm(n)
## loss of MLE
mean((theta1 - x1)^2)
## loss of oracle separable estimator
mean((theta1 - gg_rule(x1, 1, x2, 1, theta1, theta2))^2)
}
