% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmleb.R
\name{gmleb}
\alias{gmleb}
\title{General maximum likelihood empirical Bayes estimator of Jiang and Zhang (2009)}
\usage{
gmleb(x1, s1, init = NULL, grid = NULL, M = 300, tol = 1e-05, maxit = 1000)
}
\arguments{
\item{x1}{Gaussian sequence}

\item{s1}{standard deviation of Gaussian sequence}

\item{init}{initial values for masses of discrete prior, default is discrete uniform over support points}

\item{grid}{grid of support points for discrete prior, default is M equally spaced points between min(x1) and max(x1)}

\item{M}{number of grid points}

\item{tol}{error tolerance of convergence of log likelihood}

\item{maxit}{maximum number of iterations}
}
\value{
estimated values of means of primary Gaussian sequence
}
\description{
Estimates mean vector of a homoscedastic sequence of independent Gaussian observations using the nonparametric maximum likelihood procedure of Jiang and Zhang (2009)
}
\examples{
## generate data
n = 250
set.seed(1)
theta1 = rnorm(n)
x1 = theta1 + rnorm(n)
## loss of MLE
mean((theta1 - x1)^2)
## loss of GMLEB
mean((theta1 - gmleb(x1, 1)$theta_hat)^2)
}
