% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{npmle_sym_mat}
\alias{npmle_sym_mat}
\title{npmle_sym_mat}
\usage{
npmle_sym_mat(D, maxit = 200, tol = 1e-04, verbose = FALSE)
}
\arguments{
\item{D}{density matrix}

\item{maxit}{maximum number of iterations}

\item{tol}{error tolerance of convergence of log likelihood}

\item{verbose}{whether to output the error in each iteration}
}
\value{
estimated weights of grids
}
\description{
estimate the weights of grid points maximizing likelihood with given density matrix
}
