% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf0.R
\name{optimal.knots}
\alias{optimal.knots}
\title{Optimal Knots for wtf0}
\usage{
optimal.knots(x, bw)
}
\arguments{
\item{x}{Gaussian sequence}

\item{bw}{scalar bandwidth for Gaussian kernel density estimate}
}
\value{
vector of optimal knots given x and bw (one element shorter than \code{x})
}
\description{
Compute the optimal knot placement for \code{wtf0()} by minimize f_h(t) between
each consecutive pair of observations. These knots are optimal for the risk
estimate, but may not produce a better minimizer of the true risk.
}
\examples{
set.seed(1)
theta = rnorm(250)
x = theta + rnorm(250)
bw = 0.25

# optimal knots
wtf0(x, 1, bw = bw, knots = optimal.knots(x, bw))$risk.est

# approximate knots
wtf0(x, 1, bw = bw)$risk.est
}
