% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_comte}
\alias{predict_comte}
\title{Predicting function for compound decision method}
\usage{
predict_comte(comte_obj, newx)
}
\arguments{
\item{comte_obj}{returned object from comte function that contains important information for predicting.}

\item{newx}{\eqn{m x p} matrix corresponds to testing data.}
}
\value{
\item{esty}{\eqn{m x q} matrix of estimated y based on newx. }
}
\description{
The function takes returned object from comte and testing data as input. Return value is the prediction for corresponding output.
}
\examples{
\donttest{
## generate data
p = 10
q = 5
n = 50
x = matrix(rnorm(n*p,0,10), n, p)
beta = matrix(rnorm(p*q,0,10), q, p)
e = matrix(rnorm(n*q,0,0.1),n,q)
y = x \%*\% t(beta) + e
s2 = matrix(rep(0.1,q), q, 1)
## initialize parameters for EM algorithm 
x_test = matrix(rnorm(n*p,0,1), n, p)
## set minimal variance estimation min_s2 = 0.1
output1 = comte(y=y, x=x, S=beta, min_s2=0.1)
esty1 = predict_comte(output1, x_test)
## use distinct variance from multivariate linear regression models
output2 = comte(y=y, x=x, S=cbind(beta,s2))
esty2 = predict_comte(output2, x_test)
}

}
