% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf1.R
\name{TV1.oracle.highprob}
\alias{TV1.oracle.highprob}
\title{High-probability Bound on Oracle TV(d')}
\usage{
TV1.oracle.highprob(x, s)
}
\arguments{
\item{x}{Gaussian sequence}

\item{s}{standard deviation}
}
\value{
an upper bound on oracle TV(d')
}
\description{
A high-probability upper bound for the total variation of the derivative of
the oracle decision rule, d*. This is calculated by first bounding TV(d*')
using the range of unobserved means, then estimating the range by expanding
the range of the observations.
}
\examples{
theta = rnorm(250)
TV1.oracle.highprob(theta, 1)
}
