% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf1.R
\name{TV1.oracle.plugin}
\alias{TV1.oracle.plugin}
\title{Plug-in Estimate Oracle TV(d')}
\usage{
TV1.oracle.plugin(x, s, ...)
}
\arguments{
\item{x}{Gaussian sequence}

\item{s}{standard deviation}

\item{...}{other parameters passed into stats::integrate()}
}
\value{
an estimate of oracle TV(d')
}
\description{
A plug-in estimate of the total variation of the derivative of the oracle
decision rule, d*. This is calculated by plugging-in X_i for theta_i in
d*'', then R's numerical integration is used to estimate TV(d*').
}
\examples{
theta = rnorm(250)
TV1.oracle.plugin(theta, 1)
}
