% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf1.R
\name{wtf1}
\alias{wtf1}
\alias{wtf1.cvxr}
\alias{wtf1.mosek}
\title{Monotone Degree-One Weighted Trend Filter}
\usage{
wtf1(
  x,
  s,
  tau = TV1.oracle.plugin(x = x, s = s),
  M = ceiling(sqrt(length(x))),
  knots = NULL,
  backend = "CVXR",
  solver = "OSQP",
  ...
)

wtf1.cvxr(x, s, tau, knots = NULL, solver = "OSQP", ...)

wtf1.mosek(x, s, tau, knots = NULL)
}
\arguments{
\item{x}{Gaussian sequence}

\item{s}{standard deviation}

\item{tau}{a scalar bound on TV(d')}

\item{M}{number of evenly spaced knots (overrides knots if both specified)}

\item{knots}{locations of knots}

\item{backend}{one of "CVXR" or "MOSEK"}

\item{solver}{solver backend for CVXR}

\item{...}{additional parameters passed to CVXR::psolve()
(if CVXR backend is used)}
}
\value{
\item{theta_hat}{estimated values of means of Gaussian sequence}
\item{x}{original Gaussian sequence}
\item{s}{known standard deviation}
\item{SURE}{value of minimum risk estimate}
\item{tau}{the user specified bound on TV(d')}
\item{TV1}{the value of TV(d') for minimizer}
\item{intercept}{intercept of fitted estimator}
\item{slopes}{slopes of fitted estimator}
\item{knots}{location of knots}
\item{backend}{which version of wtf1 was used}
\item{solver}{solver used by CVXR}
\item{...}{additional inputed parameters}
}
\description{
Estimate the mean of a homoscedastic sequence of independent Gaussian
observations under squared error loss. The optimal separable estimator is
estimated by minimizing an unbiased estimate of the risk under a monotone
non-decreasing constraint. Under the monotonicity constraint, the unbiased
risk estimate looks like a degree-one weighted trend filter.
}
\details{
The use of the MOSEK backend or solver required a properly installed \code{Rmosek}
package. The latest installation instructions found here:
https://docs.mosek.com/latest/install/index.html and
https://docs.mosek.com/latest/rmosek/install-interface.html.

The CVXR and MOSEK backends are designed for internal use only and do not
validate inputs.
}
\examples{
# basic usage
set.seed(1)
theta = rnorm(250)
x = theta + rnorm(250)
res = wtf1(x, s = 1, tau = 1.2, knots = -2:2)
mean((theta - res$theta_hat)^2)
}
