% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{A Compound Decision Approach to Covariance Matrix Estimation}
\usage{
msg(
  X,
  K = 10,
  n_start = 20,
  centered = FALSE,
  maxit = 200,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{X}{n x p data matrix, each column is a feature}

\item{K}{number of clusters}

\item{centered}{whether to centralize each column of X}

\item{maxit}{maximum number of iterations}

\item{tol}{error tolerance of convergence of log likelihood}

\item{verbose}{whether to output the error in each iteration}
}
\value{
estimated p x p covariance matrix
}
\description{
Estimates covariance matrix for data from Gaussian distribution
}
\examples{
## generate data
n = 100; p = 30
set.seed(1)
sigma.x = matrix(0.9,p,p)
diag(sigma.x) = 1
X = MASS::mvrnorm(n,rep(0,p),sigma.x)
## loss of msg
norm(sigma.x - msg(X,K=p),'f')
}
