% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{msg}
\alias{msg}
\title{Matrix shrinkage by g-modeling method}
\usage{
msg(X, d = 10L, centered = FALSE, maxit = 200L, tol = 1e-04, verbose = FALSE)
}
\arguments{
\item{X}{\eqn{n x p} matrix of data generated from p-variate multivariate normal distribution}

\item{d}{interger. It controls the number of support points. Default value is 10.}

\item{centered}{bool. Whether X has been centered. If center=false, it is known that the multivariate
distribution has mean of zeros. Otherwise means are unknown. Default value is false.}

\item{maxit}{maximum number of allowable iterations. Default value is 200.}

\item{tol}{error tolerance for convergence of EM algorithm. Default value is 1e-04.}

\item{verbose}{bool. Whether print out error information during EM iteration. Default value is false.}
}
\value{
\itemize{
\item \code{est_cov}: \eqn{p x p} matrix of estiamted covariance matrix.
}
}
\description{
The function use shrinkage method to estimate covariance matrix with given multinormal data.
This method applies empirical method and approximate the optimal separable decision rule by
EM algorithm.
}
\examples{
\donttest{
## set parameters
p = 10
d = 20
n = 100
## generate multivariate normal data X, with mean of zeros and identity matrix as covariance matrix.
X = matrix(rnorm(n*p,0,1), n, p)
## estimate covariance with given data X
cov = msg(X=X, d=d)
}

}
